Move webhook strategy stuff into its own package
