+++ src/main/java/com/flightstats/hub/webhook/WebhookLeader.java
@@ -18,6 +18,7 @@ import com.flightstats.hub.model.ContentPath;
 import com.flightstats.hub.util.RuntimeInterruptedException;
 import com.flightstats.hub.util.Sleeper;
 import com.flightstats.hub.util.TimeUtil;
+import com.flightstats.hub.webhook.strategy.WebhookStrategy;
 import lombok.extern.slf4j.Slf4j;
 import org.joda.time.DateTime;
