+++ src/main/java/com/flightstats/hub/webhook/WebhookService.java
@@ -11,6 +11,7 @@ import com.flightstats.hub.model.ChannelConfig;
 import com.flightstats.hub.model.ContentKey;
 import com.flightstats.hub.model.ContentPath;
 import com.flightstats.hub.util.RequestUtils;
+import com.flightstats.hub.webhook.strategy.WebhookStrategy;
 import com.google.inject.name.Named;
 import lombok.extern.slf4j.Slf4j;
 import org.joda.time.DateTime;
