+++ src/main/java/com/flightstats/hub/webhook/strategy/WebhookStrategy.java
@@ -1,22 +1,22 @@
-package com.flightstats.hub.webhook;
+package com.flightstats.hub.webhook.strategy;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.flightstats.hub.cluster.ClusterCacheDao;
 import com.flightstats.hub.dao.aws.ContentRetriever;
-import com.flightstats.hub.metrics.ActiveTraces;
 import com.flightstats.hub.model.ContentKey;
 import com.flightstats.hub.model.ContentPath;
 import com.flightstats.hub.model.MinutePath;
 import com.flightstats.hub.model.SecondPath;
 import com.flightstats.hub.util.TimeUtil;
+import com.flightstats.hub.webhook.Webhook;
 
 import java.util.Optional;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-interface WebhookStrategy extends AutoCloseable {
+public interface WebhookStrategy extends AutoCloseable {
 
     static ContentPath createContentPath(Webhook webhook) {
         if (webhook.isSecond()) {
